default in rancher should be create new bridge [skip ci]
If bridge "netinstall" does not exist, a new one will be created by default and veth will be added as part of the NETINSTALL.rsc rancher script.  Note: any _bridge_ created by the script will not be removed as part of clean.  But the veth bridge _port_ member will be removed by clean.